[FIX] New directory, moving around, comicStore instead of props
