+++ skill-db/frontend/src/components/UserComponents/ChipWithInfotext.vue
