+++ skill-db/frontend/src/views/LoginView.vue
@@ -4,7 +4,7 @@
     <v-row>
 
       <v-col cols="12" class="d-lg-none">
-        <ComicOfTheDay :source="comicStore.source" :alt="comicStore.description" :mobile="true"/>
+        <ComicOfTheDay :mobile="true"/>
       </v-col>
 
       <v-col cols="12" lg="6" class="text-center mt-10">
@@ -32,9 +32,7 @@
                     label="Username"
                     v-model="userRequestLogin.username"
                     variant="outlined"
-                    :rules="[rules.required]"
-                >
-                </v-text-field>
+                    :rules="[rules.required]"/>
                 <v-text-field class="align-content-sp"
                               v-model="userRequestLogin.password"
                               :append-inner-icon="visible ? 'mdi-eye' : 'mdi-eye-off'"
@@ -44,8 +42,7 @@
                               @click:append-inner="visible = !visible"
                               :rules="[rules.required, rules.min]"
                               counter
-                              @keydown.enter="login"
-                ></v-text-field>
+                              @keydown.enter="login"/>
                 <v-btn @click="login" class="mb-10 mt-10" block
                        :disabled="!valid" elevation="0">
                   Login
@@ -57,7 +54,7 @@
       </v-col>
 
       <v-col cols="6" class="d-none d-lg-block">
-        <ComicOfTheDay :source="comicStore.source" :alt="comicStore.description" :mobile="false"/>
+        <ComicOfTheDay :mobile="false"/>
       </v-col>
 
     </v-row>
@@ -68,8 +65,8 @@
 <script setup lang="ts">
 import {useAuthStore} from "@/stores/auth";
 import {LoginRequest} from "@/models/LoginRequest";
-import ComicOfTheDay from "@/components/ComicOfTheDay.vue";
-import {useComicStore} from "@/stores/ComicStore";
+import ComicOfTheDay from "@/components/LoginComponents/ComicOfTheDay.vue";
+import {useComicStore} from "@/components/LoginComponents/ComicStore";
 import {ref} from "vue";
 
 const authStore = useAuthStore();
@@ -79,8 +76,8 @@ comicStore.loadComicOfTheDay();
 const visible = ref(false);
 const valid = ref(false);
 const rules = {
-  required: value => !!value || 'Erforderlich.',
-  min: v => v.length >= 8 || 'Min. 8 Buchstaben',
+  required: (value:string) => !!value || 'Erforderlich.',
+  min: (v:string) => v.length >= 8 || 'Min. 8 Buchstaben',
 };
 const userRequestLogin = ref(new LoginRequest());
